requirements.txt
crewai
openai
crewai[tools]
openai-agents
beautifulsoup4
faiss-cpu
sentence-transformers
numpy
